CHE-8571 add UD permissive to unknown recipe types
