+++ dashboard/src/app/workspaces/workspace-details/che-recipe.service.ts
@@ -28,7 +28,7 @@ export class CheRecipeService {
   /**
    * Default constructor that is using resource
    */
-  constructor ($log: ng.ILogService) {
+  constructor($log: ng.ILogService) {
     this.$log = $log;
   }
 
@@ -39,7 +39,7 @@ export class CheRecipeService {
    * @returns {boolean}
    */
   isScalable(recipe: che.IRecipe): boolean {
-    return this.isCompose(recipe) || this.isOpenshift(recipe);
+    return this.isCompose(recipe) || this.isKubernetes(recipe) || this.isOpenshift(recipe);
   }
 
   /**
@@ -57,14 +57,23 @@ export class CheRecipeService {
   }
 
   /**
+   * Returns true if the environment's recipe type is kubernetes.
+   *
+   * @param {che.IRecipe} recipe
+   * @returns {boolean}
+   */
+  isKubernetes(recipe: che.IRecipe): boolean {
+    return this.getRecipeType(recipe) === CheRecipeTypes.KUBERNETES;
+  }
+
+  /**
    * Returns true if the environment's recipe type is openshift.
    *
    * @param {che.IRecipe} recipe
    * @returns {boolean}
    */
   isOpenshift(recipe: che.IRecipe): boolean {
-    const recipeType = this.getRecipeType(recipe);
-    return recipeType === null ? false : recipeType === CheRecipeTypes.OPENSHIFT;
+    return this.getRecipeType(recipe) === CheRecipeTypes.OPENSHIFT;
   }
 
   /**
@@ -74,8 +83,7 @@ export class CheRecipeService {
    * @returns {boolean}
    */
   isCompose(recipe: che.IRecipe): boolean {
-    const recipeType = this.getRecipeType(recipe);
-    return recipeType === null ? false : recipeType === CheRecipeTypes.COMPOSE;
+    return this.getRecipeType(recipe) === CheRecipeTypes.COMPOSE;
   }
 
   /**
@@ -85,8 +93,7 @@ export class CheRecipeService {
    * @returns {boolean}
    */
   isDockerfile(recipe: che.IRecipe): boolean {
-    const recipeType = this.getRecipeType(recipe);
-    return recipeType === null ? false : recipeType === CheRecipeTypes.DOCKERFILE;
+    return this.getRecipeType(recipe) === CheRecipeTypes.DOCKERFILE;
   }
 
   /**
@@ -96,8 +103,7 @@ export class CheRecipeService {
    * @returns {boolean}
    */
   isDockerimage(recipe: che.IRecipe): boolean {
-    const recipeType = this.getRecipeType(recipe);
-    return recipeType === null ? false : recipeType === CheRecipeTypes.DOCKERIMAGE;
+    return this.getRecipeType(recipe) === CheRecipeTypes.DOCKERIMAGE;
   }
 
   /**
