+++ dashboard/src/app/workspaces/workspace-details/workspace-machines/edit-machine-dialog/edit-machine-dialog.html
@@ -48,6 +48,7 @@
                       editor-state="editorState"
                       validator="editMachineDialogController.isRecipeValid()"
                       on-content-change="editMachineDialogController.parseMachineRecipe()"
+                      editor-read-only="editMachineDialogController.isEditorReadOnly"
                       editor-mode="{{editMachineDialogController.editorMode}}"></che-editor>
         </div>
       </div>
