+++ dashboard/src/components/widget/editor/che-editor.controller.ts
@@ -44,6 +44,9 @@ export class CheEditorController {
    */
   private editorOptions: {
     mode?: string;
+    readOnly?: boolean;
+    lineWrapping?: boolean;
+    lineNumbers?: boolean;
     onLoad: Function;
   };
   /**
@@ -63,6 +66,10 @@ export class CheEditorController {
    */
   private onContentChange: Function;
   /**
+   * Is editor read only.
+   */
+  private editorReadOnly: boolean;
+  /**
    * Editor mode.
    */
   private editorMode: string;
@@ -77,6 +84,9 @@ export class CheEditorController {
   constructor($timeout: ng.ITimeoutService) {
     this.editorOptions = {
       mode: angular.isString(this.editorMode) ? this.editorMode : 'application/json',
+      readOnly: this.editorReadOnly ? this.editorReadOnly : false,
+      lineWrapping: true,
+      lineNumbers: true,
       onLoad: (editor: IEditor) => {
         $timeout(() => {
           editor.refresh();
